Move AddTemplate component into Pages directory
